Added config JSON file for Fleet
Since I've switched to Jetbrains Fleet as my editing software (VSC ans VS22 is better, but I like the aesthetics, so I'll be using it for a while), I needed to create a config JSON file, which compiles and runs the the compiled program, so I don't have to do it manually. The folder's name should be .fleet and the json's name should be run.json.